Evaluate Function (iteratively) at Grid of Input Arguments

PowerGrid is an apply-like function, allowing to evaluate a function at the crossings of a set of parameters. The result is saved in an array with attributes that optimize further usage by functions in package powergrid. In particular, performing a function iteratively (using parallel computing if required) is implemented conveniently. The typical use is for evaluating statistical power at a grid of assumed parameters.

Usage

PowerGrid(
  pars,
  fun,
  more_args = NULL,
  n_iter = NA,
  summarize = TRUE,
  summary_function = mean,
  parallel = FALSE,
  n_cores = future::availableCores() - 1
)

Arguments

pars

A list where each element is a vector of values named as one of the arguments of fun. fun is applied to the full grid crossing the values of each of these parameters.

fun

Function applied at each combination of pars. Arguments may contain all element names of pars and more_args. Output should be a vector, typically of length one, but more outputs are generally handled properly in further functions of powergrid.

more_args

Fixed arguments to fun that are not in pars. (internally used in .mapply for supplying argument MoreArgs)

n_iter

If not NA, function fun is applied n_iter times at each point in the grid defined by pars.

summarize

Logical indicating whether iterations (if n_iter is given) are to be summarized by summary_function.

summary_function

A function to be applied to aggregate across simulations. Defaults to mean, ignored when keep_sims == TRUE or when is.na(n_iter).

parallel

Logical indicating whether parallel computing should be applied. If TRUE, future::future_replicate is used internally.

n_cores

Passed on to future_replicate

Value

An array of class "power_array"

Details

Function fun is evaluated at each combination of the argument values listed in pars and its results are stored in an array whose dimensions (and dimnames()) are defined by pars. If function fun returns more than one value (always as a single vector), the array will have an additional dimension, with levels named after the names of the returned vector (if given). If n_iter is not NA, function fun is evaluated n_iter times. This will add to the resulting array an additional dimension 'sim'.

The elements of pars must match the argument names of fun. If input parameters are not to be part of the grid, but rather further settings, these can be passed on to fun through the argument more_args as a list with names reflecting the arguments of fun to be set.

Author

Gilles Dutilh

Examples

## ============================================
## most basic case, power function available:
## ============================================
sse_pars = list(
  n = seq(from = 10, to = 60, by = 2),
  delta = seq(from = 0.5, to = 1.5, by = 0.2), ## effect size
  sd = seq(.1, .9, .2)) ## Standard deviation
PowFun <- function(n, delta, sd){
  ptt = power.t.test(n = n/2, delta = delta, sd = sd,
                     sig.level = 0.05)
  return(ptt$power)
}
power_array = PowerGrid(pars = sse_pars, fun = PowFun, n_iter = NA)
##'
summary(power_array)
#>  Object of class: power_array
#> 
#>  Range of values: [0.12, 1] 
#>  Evaluated at:
#>       n 10, 12, 14, 16, 18, 20, 22, 24, 26, 28, 30, 32,
#>       n 34, 36, 38, 40, 42, 44, 46, 48, 50, 52, 54, 56,
#>       n 58, 60
#>   delta 0.5, 0.7, 0.9, 1.1, 1.3, 1.5
#>      sd 0.1, 0.3, 0.5, 0.7, 0.9

## ============================================
## Multiple outputs are automatically handled
## ============================================
TwoValuesFun <- function(n, delta, sd){
  p5 = power.t.test(n = n, delta = delta, sd = sd, sig.level = .05)$power
  p1 = power.t.test(n = n, delta = delta, sd = sd, sig.level = .01)$power
  return(c('p5' = p5, 'p1' = p1))
}
##
sse_pars = list(
  n = seq(from = 10, to = 60, by = 2),
  delta = seq(from = 0.5, to = 1.5, by = 0.2),
  sd = seq(.5, 1.5, .2))
array_two_returns = PowerGrid(sse_pars, TwoValuesFun)
## multiple outputs result in an additional dimension:
dimnames(array_two_returns)
#> $n
#>  [1] "10" "12" "14" "16" "18" "20" "22" "24" "26" "28" "30" "32" "34" "36" "38"
#> [16] "40" "42" "44" "46" "48" "50" "52" "54" "56" "58" "60"
#> 
#> $delta
#> [1] "0.5" "0.7" "0.9" "1.1" "1.3" "1.5"
#> 
#> $sd
#> [1] "0.5" "0.7" "0.9" "1.1" "1.3" "1.5"
#> 
#> $fun_out
#> [1] "p5" "p1"
#> 
summary(array_two_returns)
#>  Object of class: power_array
#> 
#>  Range of values: 
#>       p5: [0.1, 1]
#>       p1: [0.03, 1] 
#>  Evaluated at:
#>       n 10, 12, 14, 16, 18, 20, 22, 24, 26, 28, 30, 32,
#>       n 34, 36, 38, 40, 42, 44, 46, 48, 50, 52, 54, 56,
#>       n 58, 60
#>   delta 0.5, 0.7, 0.9, 1.1, 1.3, 1.5
#>      sd 0.5, 0.7, 0.9, 1.1, 1.3, 1.5

## ============================================
## Simulations over iterations
## ============================================
PowFun <- function(n, delta, sd){
  x1 = rnorm(n = n/2, sd = sd)
  x2 = rnorm(n = n/2, mean = delta, sd = sd)
  t.test(x1, x2)$p.value < .05
}
sse_pars = list(
  n = seq(from = 10, to = 60, by = 5),
  delta = seq(from = 0.5, to = 1.5, by = 0.2),
  sd = seq(.5, 1.5, .2))
##
n_iter = 20
power_array = PowerGrid(pars = sse_pars, fun = PowFun,
                       n_iter = n_iter)
dimnames(power_array)
#> $n
#>  [1] "10" "15" "20" "25" "30" "35" "40" "45" "50" "55" "60"
#> 
#> $delta
#> [1] "0.5" "0.7" "0.9" "1.1" "1.3" "1.5"
#> 
#> $sd
#> [1] "0.5" "0.7" "0.9" "1.1" "1.3" "1.5"
#> 
summary(power_array)
#>  Object of class: power_array
#> Containing summary over 20 iterations,
#> summarized by function `mean` (for function
#> definition, see attribute `summary_function`).
#>  Range of values: [0.05, 1] 
#>  Evaluated at:
#>       n 10, 15, 20, 25, 30, 35, 40, 45, 50, 55, 60
#>   delta 0.5, 0.7, 0.9, 1.1, 1.3, 1.5
#>      sd 0.5, 0.7, 0.9, 1.1, 1.3, 1.5