Wrap text in printed output.

Wrap text in printed output. This function is created for convenience and also so that the width of wrapping can be set by simply changing the default here. Note that this give a leading white space to each resulting line. The is done to account for the fact that wrapping deletes the whites paces, and it actually looks good.

Usage

PrintWrap(x, width = 48, ...)

Arguments

x

A single character string

width

The desired width

...

Further arguments passed to 'strwrap()'

Value

a character string with returns placed such to achieve the desired width

Author

Gilles Dutilh