indexing with [ ] for class power_array [ ]: R:%20

Method for indexing [] of objects of class power_array. The method makes sure that the resulting array is of class power_array and keeps and updates the object's attributes. These attributes are needed for various functions in the powergrid package to work well. ##' The indexing functions as normal indexing, but note that drop is FALSE by default, so that the resulting array has the same dimensions as the original array. The number of levels at each dimension may be reduced, however. ##'

Usage

# S3 method for class 'power_array'
x[..., drop = TRUE]

Arguments

x

object

...

index

drop

drop

Value

An array of class power_grid

Author

Gilles Dutilh